move occgrid -> cv image conversion to utils
