+++ map_service/src/map_server.cpp
@@ -1,5 +1,10 @@
 #include "map_service/map_server.hpp"
 
+#include <grid_map_cv/GridMapCvConverter.hpp>
+#include <grid_map_ros/GridMapRosConverter.hpp>
+
+#include "map_service/utils.hpp"
+
 namespace cad2cav {
 namespace map_service {
 
@@ -189,38 +194,8 @@ cad2cav_msgs::LandmarkDetectionList MapServer::composeWindowLandmarkMsg(
   return window_msg;
 }
 
-namespace {
-cv::Mat transposeMapImage(const cv::Mat& map_image_original) {
-  cv::Mat map_image{map_image_original.cols, map_image_original.rows, CV_8UC1},
-      temp1{map_image_original.rows, map_image_original.cols, CV_8UC1},
-      temp2{map_image_original.rows, map_image_original.cols, CV_8UC1};
-
-  cv::flip(map_image_original, temp1, 0);
-  cv::flip(temp1, temp2, 1);
-  cv::transpose(temp2, map_image);
-
-  cv::Mat map_image_colored;
-  cv::cvtColor(map_image, map_image_colored, cv::COLOR_GRAY2BGR);
-
-  return map_image_colored;
-}
-}  // namespace
-
 cv::Mat MapServer::toCvImage() const {
-  static const std::string map_layer = "base";
-
-  grid_map::GridMap grid_map;
-  if (!grid_map::GridMapRosConverter::fromOccupancyGrid(map_, map_layer,
-                                                        grid_map)) {
-    ROS_ERROR("Conversion to GridMap failed");
-  }
-
-  cv::Mat map_img;
-  if (!grid_map::GridMapCvConverter::toImage<unsigned char, 1>(
-          grid_map, map_layer, CV_8UC1, 0, 100, map_img)) {
-    ROS_ERROR("Conversion to cv::Mat failed");
-  }
-  return transposeMapImage(map_img);
+  return utils::occupancyGridToCvImage(map_);
 }
 
 }  // namespace map_service
